[DEBT-043-D8] Split execution/executor.py into focused modules
- Split executor.py (637 LoC) into 4 new modules:
  - _protocols.py (49 LoC): Protocol definitions for providers
  - _checks.py (307 LoC): Individual safety check implementations
  - _orchestration.py (189 LoC): Config classes and run_live_checks
  - _executor.py (339 LoC): TradeExecutor class
- Updated __init__.py to re-export from new modules
- All files now ≤400 lines (largest: _executor.py at 339)
- Moved type-only imports into TYPE_CHECKING blocks per TC001
- Quality gates: pre-commit, pytest (1052 passed)
