+++ src/kalshi_research/execution/__init__.py
@@ -1,5 +1,18 @@
 """Trading execution safety harness (safe-by-default)."""
 
-from kalshi_research.execution.executor import TradeExecutor, TradeSafetyError
+from kalshi_research.execution._executor import TradeExecutor, TradeSafetyError
+from kalshi_research.execution._protocols import (
+    DailyBudgetTracker,
+    MarketProvider,
+    OrderbookProvider,
+    PositionProvider,
+)
 
-__all__ = ["TradeExecutor", "TradeSafetyError"]
+__all__ = [
+    "DailyBudgetTracker",
+    "MarketProvider",
+    "OrderbookProvider",
+    "PositionProvider",
+    "TradeExecutor",
+    "TradeSafetyError",
+]
